requirements.txt
torch
numpy
sympy
pandas
scikit-learn
matplotlib
tqdm
pyyaml
tensorboard
pytest
